Document uncurried conversion through test cases. (#58)
Unfortunately there's a bug in the Reason parser.
f(. a) will parse the @bs attribute in the wrong place, when it's
an expression at the module level: https://github.com/facebook/reason/issues/2565

The new Res syntax has the correct behaviour. Unfortunately we can't
fix bugs from Reason, we would change the semantics of the program.