Refactor parsing of string template literals.
The previous approach put the scanner in "Template" mode when the parser
was going to parse template literals. This resulted in some very awkward
code; upon scanning a token there was logic checking whether we were
in "Template" mode. Every non-template literal token would pay for this
extra branch…

The new parsing strategy works differently: when the parser needs to
parse template literals, it just asks the scanner immediately for a
template literal token. This is both more performant an easier to reason
about. Template literals are a different language, it makes sense to
split this into a separate scanning function.