Report when error is missing in comma delimited regions.
When a comma is missing between nodes in a region and the current token looks like the start of something valid in the current region, we should report an error with regards to the missing comma.

Example:
  type student<'extraInfo> = {
    name: string,
    age: int
    otherInfo: 'extraInfo
  }
There is a missing comma between `int` and `otherInfo`.
`otherInfo` looks like a valid start of the record declaration.
We report the error here and then continue parsing the region.

Fixes https://github.com/rescript-lang/syntax/issues/100