Implement intelligent breaking for chained or-patterns.
If you look at the `B | C | D` patterns in the following examples:
```
switch a {
| B
| C
| D => 1
| E => 2
}
switch a {
| B | C | D => 1
| E => 2
}
```
they have two different layouts:
* In the first example the user wrote them over multiple lines.
* The second has them on one line.

This commit implements a new strategy to print these two examples.
We now look at the style of the author: did he write it over multiple lines or not?
* If it is on one line, keep it on one line (unless it breaks the column width).
* If it is written over multiple lines, force break it over multiple lines.

This is also consistent with how we print variant declaration.