Improve printing of multiline strings and template literals.
When the content of the string exceeds the line length, the printer breaks the layout. Line breaks in strings itself are just counted as characters now. However they form actual line breaks in the layout. A new type of linebreak `LiteralLine` is introduced to fix it. It is always included and forces breaks in the parents, but doesn't increase the indentation of the subsequent line.