Enable stack trace for binary; remove an `exit 1`
This makes the perf slightly slower (but basically unnoticeable), and the compilation slower, but this repo's files are compiled with stack trace enabled inside the compiler project anyway, so this is more realistic.

On the other hand, having a general try catch in res_cli that swallowed important exceptions such as `String.sub` and `Array.get` would have made debugging much harder.