Do not parse past Eof
Parsing past Eof can lead to infinite loops, as it violates the assumptions of the termination checker.
See: https://github.com/rescript-lang/syntax/pull/540

This PR makes `Parser.next` assert false when called on Eof. This should not happen as one should check the token before calling `Parser.next`.

The one exception is during lookahead, for which we provide a `nextUnsafe` function which does not make progress.