Audit of the code that might raise Eof by calling Parser.next.
There are 2 functions that the parser can call:
- next can raise Eof
- nextUnsafe does not raise

The former is counted in the progress measure. It cannot be part of infinite loops, as eventually Eof would be reached and the exception thrown.

This means that every use of Parser.next needs to be audited for exceptions, since infinite loops have been turned into runtime exceptions.
Sometimes it's enough to use Parser.expect if we know exactly what is the current token.
Sometimes the current token has just been checked, so a `@doesNotRaise` annotation is added.
Sometimes there are several call-sites to the current function that have different invariants, and this is where some changes were required.