Remove printing of arrow functions with "multiple" dots
Arrow functions with multiple dots `(. a, . b) => a + b` are considered an anti-pattern.
This syntax *will be removed* in the future. As a first step, the printer will know print this as multiple arrow functions. (Like in JavaScript)

**Before**
```rescript
let f = (. a, . b) => a + b
type f = (. int, . int) => int
```

**After**
```rescript
let f = (. a) => (. b) => a + b
type f = (. int) => (. int) => int
```